Add python/lsst/ts/report/generator/participant.py module.
This module defines the `Participant` class, which provides a way to describe a component that "participates" in an activity driven by an `Actor`.

A `Participant` will either be directly receiving commands from the `Actor` or acting as a result of something it did.
The report will show all commands the actor sent to the participants and the selected events/telemetry sent by the participants.